WIP - working on background spacebux generation
Need to deserialize planet xml to update player profile with known
planets (currently the only way to own planets is to send a ship in
system view).

Commented out section of spacebux gathering -> that'll update the last
collected time, and once that's in place, background generation on
planets you own can work.

Going to be disabling resource generation on planets you don't own
eventually.

+++ Assets/Scripts/Spaceobjects/Planet.cs
@@ -1,5 +1,6 @@
 ﻿using UnityEngine;
 using System.Collections;
+using System;
 using Superbest_random;
 
 public class Planet : PlanetaryBody {
@@ -37,7 +38,7 @@ public class Planet : PlanetaryBody {
     // Gameplay variables
     public double energyModifier;
     public double populationRate;
-    public uint population;
+    public long population;
 
 	public LineRenderer orbitPath;
 	public GameObject orbitParent;
